Coding Tracker 0.4
Added new validation, optimized part of the code, introduced Spectre.Console, different changes and fixes

+++ Main.cs
@@ -7,6 +7,7 @@ using System.IO;
 
 using System.Configuration;
 using System.Collections.Specialized;
+using static System.Runtime.InteropServices.JavaScript.JSType;
 
 var userInput = new UserInput();
 
@@ -30,67 +31,52 @@ connection.Execute(createTableQuery);
 
 while (end)
 {
-    AnsiConsole.Markup(@"[blue]Welcome to the Coding Tracker![/]
-                    
-Please, choose an option from the list below:
-1. Start a Coding Session.
-2. Add a Coding Session.
-3. Change an existed Coding Session.
-4. Show your Coding Sessions.
-5. Delete a Coding Session.
-6. Exit
+    var panelHeader = new Panel("[blue]Welcome to the Coding Tracker![/]").Padding(38,3,3,3).Expand();
+    
+    AnsiConsole.Write(panelHeader);
 
-");
+    string fruit = AnsiConsole.Prompt(
+        new SelectionPrompt<string>()
+            .Title("[blue]Please, choose an option from the list below:[/]")
+            .PageSize(10)
+            .MoreChoicesText("[grey](Move up and down to reveal more fruits)[/]")
+            .AddChoices(new[] {
+                "Start a Coding Session.", "Add a Coding Session.", "Change an existed Coding Session.",
+                "Show your Coding Sessions.", "Delete a Coding Session.", "Exit",
+                }));
     // ADD VALIDATION
-
-    Console.Write("Type your choice: ");
-    switch (userInput.CreateInt())
+    switch (fruit)
     {
-        case 1:
-            var dateInputC = DateTime.Now;
-            session.StartTime = DateTime.Parse(dateInputC.ToShortDateString() + " " + dateInputC.ToLongTimeString());
-            controller.SetTimer();
-
-            Console.WriteLine("Press any key to stop the timer.");
-            Console.ReadKey();
-
-            controller.StopTimer();
-
-            dateInputC = DateTime.Now;
-            session.EndTime = DateTime.Parse(dateInputC.ToShortDateString() + " " + dateInputC.ToLongTimeString());
-            controller.Create(connection, session);
+        case "Start a Coding Session.":
+            controller.CreateWithTimer(connection, session);
             break;
-        case 2:
-            Console.Write("Type date of your Coding Session or type T for today's date: ");
-            var dateAddInput = userInput.CreateRegex(@"^([0-2][0-9]|3[01])\.(0[1-9]|1[0-2])\.(\d{4})$|^T|t$",
-                "Wrong data format, try again. Example: 01.01.2001 or T for today's date: ");
+        case "Add a Coding Session.":
+            var dateAddInput = userInput.CreateRegex(@"^([0-2][0-9]|3[01])\.(0[1-9]|1[0-2])\.(\d{4})$|^T|t$", "Type date of your Coding Session or type T for today's date: "
+                ,"Wrong data format, try again. Example: 01.01.2001 or T for today's date: ");
             dateAddInput = UserInput.CheckT(dateAddInput);
 
-            Console.Write("Type your Coding Session start time: ");
-            var startAddTimeInput = userInput.CreateRegex(@"^([0-1][0-9]|2[0-3])\:([0-5][0-9])\:([0-5][0-9])$",
-                "Wrong data format, try again. Example: 10:10:10: ");
+            var startAddTimeInput = userInput.CreateRegex(@"^([0-1][0-9]|2[0-3])\:([0-5][0-9])\:([0-5][0-9])$", "Type your Coding Session start time: "
+                , "Wrong data format, try again. Example: 10:10:10: ");
 
-            Console.Write("Add end time of your Coding Session or type N for the time at this moment: ");
-            var endAddTimeInput = userInput.CreateRegex(@"^([0-1][0-9]|2[0-3])\:([0-5][0-9])\:([0-5][0-9])$|^N|n$",
-                "Wrong data format, try again. Example: 12:12:12 or N for the time at this moment: ");
+            var endAddTimeInput = userInput.CreateRegex(@"^([0-1][0-9]|2[0-3])\:([0-5][0-9])\:([0-5][0-9])$|^N|n$", "Add end time of your Coding Session or type N for the time at this moment: "
+                , "Wrong data format, try again. Example: 12:12:12 or N for the time at this moment: ");
 
             endAddTimeInput = UserInput.CheckN(endAddTimeInput);
             session.StartTime = DateTime.Parse(dateAddInput + " " + startAddTimeInput);
             session.EndTime = DateTime.Parse(dateAddInput + " " + endAddTimeInput);
             controller.Create(connection, session);
             break;
-        case 3:
-            Console.Write("Type date of your Coding Session or type T for today's date: ");
-            var dateChangeInput = userInput.CreateRegex(@"^([0-2][0-9]|3[01])\.(0[1-9]|1[0-2])\.(\d{4})$|^T|t$",
+        case "Change an existed Coding Session.":
+            var dateChangeInput = userInput.CreateRegex(@"^([0-2][0-9]|3[01])\.(0[1-9]|1[0-2])\.(\d{4})$|^T|t$", "Type date of your Coding Session or type T for today's date: ",
                 "Wrong data format, try again. Example: 01.01.2001 or T for today's date: ");
             dateChangeInput = UserInput.CheckT(dateChangeInput);
 
-            Console.Write("Type S to change Start Time of Session or type E to change End Time: ");
-            var sessionTimeInput = userInput.CreateRegex(@"^S|s$|^E|e$", "Please, type only listed symbols: ");
+            var sessionTimeInput = userInput.CreateRegex(@"^S|s$|^E|e$", "Type S to change Start Time of Session or type E to change End Time: ","Please, type only listed symbols: ");
 
-            var chooseChangeSession = userInput.ChooseSession(controller.Read(connection, dateChangeInput));
-            Console.Write("Type your new Coding time: ");
-            var ChangeTimeInput = userInput.CreateRegex(@"^([0-1][0-9]|2[0-3])\:([0-5][0-9])\:([0-5][0-9])$|^N|n$",
+            var addSessionList = controller.Read(connection, dateChangeInput);
+            if (addSessionList.Count == 0) break;
+            var chooseChangeSession = userInput.ChooseSession(addSessionList);
+            var ChangeTimeInput = userInput.CreateRegex(@"^([0-1][0-9]|2[0-3])\:([0-5][0-9])\:([0-5][0-9])$|^N|n$", "Type your new Coding time: ",
                 "Wrong data format, try again. Example: 10:10:10: ");
 
             if (sessionTimeInput is "S" or "s")
@@ -105,19 +91,20 @@ Please, choose an option from the list below:
             }
 
             break;
-        case 4:
+        case "Show your Coding Sessions.":
             controller.Read(connection);
             break;
-        case 5:
-            Console.Write("Type data of your Coding Session you chose to delete or type T for today's date: ");
-            var deleteDateInput = userInput.CreateRegex(@"^([0-2][0-9]|3[01])\.(0[1-9]|1[0-2])\.(\d{4})$|^T|t$",
+        case "Delete a Coding Session.":
+            var deleteDateInput = userInput.CreateRegex(@"^([0-2][0-9]|3[01])\.(0[1-9]|1[0-2])\.(\d{4})$|^T|t$", "Type data of your Coding Session you chose to delete or type T for today's date: ",
                 "Wrong data format, try again. Example: 01.01.2001 or T for today's date: ");
             deleteDateInput = UserInput.CheckT(deleteDateInput);
 
-            var chooseDeleteSession = userInput.ChooseSession(controller.Read(connection, deleteDateInput));
+            var deleteSessionList = controller.Read(connection, deleteDateInput);
+            if (deleteSessionList.Count == 0) break;
+            var chooseDeleteSession = userInput.ChooseSession(deleteSessionList);
             controller.Delete(connection, deleteDateInput, chooseDeleteSession.StartTime.ToLongTimeString());
             break;
-        case 6:
+        case "Exit":
             end = false;
             break;
         default:
@@ -125,7 +112,7 @@ Please, choose an option from the list below:
             break;
     }
 
-    Console.WriteLine("Press any key to continue...");
+    Console.WriteLine("\nPress any key to continue...");
     Console.ReadKey();
     Console.Clear();
 }
